requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.115.13
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
python-slugify==8.0.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
